сделала пагинацию в одном компоненте без функции в отдельном файле
